+++ work/src/org/kuali/kfs/module/bc/businessobject/BudgetConstructionAppointmentFundingLoad.java
@@ -25,6 +25,7 @@
 
 package org.kuali.module.budget.bo;
 
+import java.math.BigDecimal;
 import java.util.LinkedHashMap;
 
 import org.kuali.core.bo.BusinessObjectBase;
@@ -48,14 +49,14 @@ public class BudgetConstructionAppointmentFundingLoad extends BusinessObjectBase
 	private String emplid;
 	private String appointmentFundingDurationCode;
 	private KualiDecimal appointmentRequestedCsfAmount;
-	private Integer appointmentRequestedCsfFteQuantity;
-	private Integer appointmentRequestedCsfTimePercent;
+	private BigDecimal appointmentRequestedCsfFteQuantity;
+	private BigDecimal appointmentRequestedCsfTimePercent;
 	private KualiDecimal appointmentTotalIntendedAmount;
-	private Integer appointmentTotalIntendedFteQuantity;
+	private BigDecimal appointmentTotalIntendedFteQuantity;
 	private KualiDecimal appointmentRequestedAmount;
-	private Integer appointmentRequestedTimePercent;
-	private Integer appointmentRequestedFteQuantity;
-	private Integer appointmentRequestedPayRate;
+	private BigDecimal appointmentRequestedTimePercent;
+	private BigDecimal appointmentRequestedFteQuantity;
+	private BigDecimal appointmentRequestedPayRate;
 	private String appointmentFundingDeleteCode;
 	private Integer appointmentFundingMonth;
 
@@ -286,7 +287,7 @@ public class BudgetConstructionAppointmentFundingLoad extends BusinessObjectBase
 	 * @return - Returns the appointmentRequestedCsfFteQuantity
 	 * 
 	 */
-	public Integer getAppointmentRequestedCsfFteQuantity() { 
+	public BigDecimal getAppointmentRequestedCsfFteQuantity() { 
 		return appointmentRequestedCsfFteQuantity;
 	}
 
@@ -296,7 +297,7 @@ public class BudgetConstructionAppointmentFundingLoad extends BusinessObjectBase
 	 * @param - appointmentRequestedCsfFteQuantity The appointmentRequestedCsfFteQuantity to set.
 	 * 
 	 */
-	public void setAppointmentRequestedCsfFteQuantity(Integer appointmentRequestedCsfFteQuantity) {
+	public void setAppointmentRequestedCsfFteQuantity(BigDecimal appointmentRequestedCsfFteQuantity) {
 		this.appointmentRequestedCsfFteQuantity = appointmentRequestedCsfFteQuantity;
 	}
 
@@ -307,7 +308,7 @@ public class BudgetConstructionAppointmentFundingLoad extends BusinessObjectBase
 	 * @return - Returns the appointmentRequestedCsfTimePercent
 	 * 
 	 */
-	public Integer getAppointmentRequestedCsfTimePercent() { 
+	public BigDecimal getAppointmentRequestedCsfTimePercent() { 
 		return appointmentRequestedCsfTimePercent;
 	}
 
@@ -317,7 +318,7 @@ public class BudgetConstructionAppointmentFundingLoad extends BusinessObjectBase
 	 * @param - appointmentRequestedCsfTimePercent The appointmentRequestedCsfTimePercent to set.
 	 * 
 	 */
-	public void setAppointmentRequestedCsfTimePercent(Integer appointmentRequestedCsfTimePercent) {
+	public void setAppointmentRequestedCsfTimePercent(BigDecimal appointmentRequestedCsfTimePercent) {
 		this.appointmentRequestedCsfTimePercent = appointmentRequestedCsfTimePercent;
 	}
 
@@ -349,7 +350,7 @@ public class BudgetConstructionAppointmentFundingLoad extends BusinessObjectBase
 	 * @return - Returns the appointmentTotalIntendedFteQuantity
 	 * 
 	 */
-	public Integer getAppointmentTotalIntendedFteQuantity() { 
+	public BigDecimal getAppointmentTotalIntendedFteQuantity() { 
 		return appointmentTotalIntendedFteQuantity;
 	}
 
@@ -359,7 +360,7 @@ public class BudgetConstructionAppointmentFundingLoad extends BusinessObjectBase
 	 * @param - appointmentTotalIntendedFteQuantity The appointmentTotalIntendedFteQuantity to set.
 	 * 
 	 */
-	public void setAppointmentTotalIntendedFteQuantity(Integer appointmentTotalIntendedFteQuantity) {
+	public void setAppointmentTotalIntendedFteQuantity(BigDecimal appointmentTotalIntendedFteQuantity) {
 		this.appointmentTotalIntendedFteQuantity = appointmentTotalIntendedFteQuantity;
 	}
 
@@ -391,7 +392,7 @@ public class BudgetConstructionAppointmentFundingLoad extends BusinessObjectBase
 	 * @return - Returns the appointmentRequestedTimePercent
 	 * 
 	 */
-	public Integer getAppointmentRequestedTimePercent() { 
+	public BigDecimal getAppointmentRequestedTimePercent() { 
 		return appointmentRequestedTimePercent;
 	}
 
@@ -401,7 +402,7 @@ public class BudgetConstructionAppointmentFundingLoad extends BusinessObjectBase
 	 * @param - appointmentRequestedTimePercent The appointmentRequestedTimePercent to set.
 	 * 
 	 */
-	public void setAppointmentRequestedTimePercent(Integer appointmentRequestedTimePercent) {
+	public void setAppointmentRequestedTimePercent(BigDecimal appointmentRequestedTimePercent) {
 		this.appointmentRequestedTimePercent = appointmentRequestedTimePercent;
 	}
 
@@ -412,7 +413,7 @@ public class BudgetConstructionAppointmentFundingLoad extends BusinessObjectBase
 	 * @return - Returns the appointmentRequestedFteQuantity
 	 * 
 	 */
-	public Integer getAppointmentRequestedFteQuantity() { 
+	public BigDecimal getAppointmentRequestedFteQuantity() { 
 		return appointmentRequestedFteQuantity;
 	}
 
@@ -422,7 +423,7 @@ public class BudgetConstructionAppointmentFundingLoad extends BusinessObjectBase
 	 * @param - appointmentRequestedFteQuantity The appointmentRequestedFteQuantity to set.
 	 * 
 	 */
-	public void setAppointmentRequestedFteQuantity(Integer appointmentRequestedFteQuantity) {
+	public void setAppointmentRequestedFteQuantity(BigDecimal appointmentRequestedFteQuantity) {
 		this.appointmentRequestedFteQuantity = appointmentRequestedFteQuantity;
 	}
 
@@ -433,7 +434,7 @@ public class BudgetConstructionAppointmentFundingLoad extends BusinessObjectBase
 	 * @return - Returns the appointmentRequestedPayRate
 	 * 
 	 */
-	public Integer getAppointmentRequestedPayRate() { 
+	public BigDecimal getAppointmentRequestedPayRate() { 
 		return appointmentRequestedPayRate;
 	}
 
@@ -443,7 +444,7 @@ public class BudgetConstructionAppointmentFundingLoad extends BusinessObjectBase
 	 * @param - appointmentRequestedPayRate The appointmentRequestedPayRate to set.
 	 * 
 	 */
-	public void setAppointmentRequestedPayRate(Integer appointmentRequestedPayRate) {
+	public void setAppointmentRequestedPayRate(BigDecimal appointmentRequestedPayRate) {
 		this.appointmentRequestedPayRate = appointmentRequestedPayRate;
 	}
