+++ work/src/org/kuali/kfs/module/bc/businessobject/BudgetConstructionBalanceByAccount.java
@@ -25,6 +25,7 @@
 
 package org.kuali.module.budget.bo;
 
+import java.math.BigDecimal;
 import java.util.LinkedHashMap;
 
 import org.kuali.core.bo.BusinessObjectBase;
@@ -54,12 +55,12 @@ public class BudgetConstructionBalanceByAccount extends BusinessObjectBase {
 	private String typeFinancialReportSortCode;
 	private String financialConsolidationSortCode;
 	private String levelFinancialReportSortCode;
-	private Integer appointmentRequestedFteQuantity;
-	private Integer appointmentRequestedCsfFteQuantity;
-	private Integer csfFullTimeEmploymentQuantity;
+	private BigDecimal appointmentRequestedFteQuantity;
+	private BigDecimal appointmentRequestedCsfFteQuantity;
+	private BigDecimal csfFullTimeEmploymentQuantity;
 	private KualiDecimal accountLineAnnualBalanceAmount;
 	private KualiDecimal financialBeginningBalanceLineAmount;
-	private Integer positionCsfLeaveFteQuantity;
+	private BigDecimal positionCsfLeaveFteQuantity;
 
     private ObjectCode financialObject;
 	private Account account;
@@ -334,7 +335,7 @@ public class BudgetConstructionBalanceByAccount extends BusinessObjectBase {
 	 * @return - Returns the appointmentRequestedFteQuantity
 	 * 
 	 */
-	public Integer getAppointmentRequestedFteQuantity() { 
+	public BigDecimal getAppointmentRequestedFteQuantity() { 
 		return appointmentRequestedFteQuantity;
 	}
 
@@ -344,7 +345,7 @@ public class BudgetConstructionBalanceByAccount extends BusinessObjectBase {
 	 * @param - appointmentRequestedFteQuantity The appointmentRequestedFteQuantity to set.
 	 * 
 	 */
-	public void setAppointmentRequestedFteQuantity(Integer appointmentRequestedFteQuantity) {
+	public void setAppointmentRequestedFteQuantity(BigDecimal appointmentRequestedFteQuantity) {
 		this.appointmentRequestedFteQuantity = appointmentRequestedFteQuantity;
 	}
 
@@ -355,7 +356,7 @@ public class BudgetConstructionBalanceByAccount extends BusinessObjectBase {
 	 * @return - Returns the appointmentRequestedCsfFteQuantity
 	 * 
 	 */
-	public Integer getAppointmentRequestedCsfFteQuantity() { 
+	public BigDecimal getAppointmentRequestedCsfFteQuantity() { 
 		return appointmentRequestedCsfFteQuantity;
 	}
 
@@ -365,7 +366,7 @@ public class BudgetConstructionBalanceByAccount extends BusinessObjectBase {
 	 * @param - appointmentRequestedCsfFteQuantity The appointmentRequestedCsfFteQuantity to set.
 	 * 
 	 */
-	public void setAppointmentRequestedCsfFteQuantity(Integer appointmentRequestedCsfFteQuantity) {
+	public void setAppointmentRequestedCsfFteQuantity(BigDecimal appointmentRequestedCsfFteQuantity) {
 		this.appointmentRequestedCsfFteQuantity = appointmentRequestedCsfFteQuantity;
 	}
 
@@ -376,7 +377,7 @@ public class BudgetConstructionBalanceByAccount extends BusinessObjectBase {
 	 * @return - Returns the csfFullTimeEmploymentQuantity
 	 * 
 	 */
-	public Integer getCsfFullTimeEmploymentQuantity() { 
+	public BigDecimal getCsfFullTimeEmploymentQuantity() { 
 		return csfFullTimeEmploymentQuantity;
 	}
 
@@ -386,7 +387,7 @@ public class BudgetConstructionBalanceByAccount extends BusinessObjectBase {
 	 * @param - csfFullTimeEmploymentQuantity The csfFullTimeEmploymentQuantity to set.
 	 * 
 	 */
-	public void setCsfFullTimeEmploymentQuantity(Integer csfFullTimeEmploymentQuantity) {
+	public void setCsfFullTimeEmploymentQuantity(BigDecimal csfFullTimeEmploymentQuantity) {
 		this.csfFullTimeEmploymentQuantity = csfFullTimeEmploymentQuantity;
 	}
 
@@ -439,7 +440,7 @@ public class BudgetConstructionBalanceByAccount extends BusinessObjectBase {
 	 * @return - Returns the positionCsfLeaveFteQuantity
 	 * 
 	 */
-	public Integer getPositionCsfLeaveFteQuantity() { 
+	public BigDecimal getPositionCsfLeaveFteQuantity() { 
 		return positionCsfLeaveFteQuantity;
 	}
 
@@ -449,7 +450,7 @@ public class BudgetConstructionBalanceByAccount extends BusinessObjectBase {
 	 * @param - positionCsfLeaveFteQuantity The positionCsfLeaveFteQuantity to set.
 	 * 
 	 */
-	public void setPositionCsfLeaveFteQuantity(Integer positionCsfLeaveFteQuantity) {
+	public void setPositionCsfLeaveFteQuantity(BigDecimal positionCsfLeaveFteQuantity) {
 		this.positionCsfLeaveFteQuantity = positionCsfLeaveFteQuantity;
 	}
